feat: first endpoint

+++ src/main/java/org/circle8/controller/TransporteController.java
@@ -16,9 +16,9 @@ import io.javalin.http.Context;
 @Singleton
 public class TransporteController {
 	private final static TransporteResponse mock = TransporteResponse.builder()
-		.id(1)
+		.id(1L)
 		.fechaAcordada(ZonedDateTime.of(2023, 1, 1, 16, 30, 0, 0, Dates.UTC))
-		.transaccionId(1)
+		.transaccionId(1L)
 		.transportistaUri("/transportista/1")
 		.build();
 
@@ -28,7 +28,7 @@ public class TransporteController {
 	public ApiResponse list(Context ctx) {
 		final var l = List.of(
 			mock,
-			mock.toBuilder().id(2).build()
+			mock.toBuilder().id(2L).build()
 		);
 
 		return new ListResponse<>(0, 1, 2, null, null, l);
@@ -38,7 +38,7 @@ public class TransporteController {
 	 * GET /transporte/{id}
 	 */
 	public ApiResponse get(Context ctx) {
-		return mock.toBuilder().id(Integer.parseInt(ctx.pathParam("id"))).build();
+		return mock.toBuilder().id(Long.parseLong(ctx.pathParam("id"))).build();
 	}
 
 	/**
@@ -46,7 +46,7 @@ public class TransporteController {
 	 */
 	public ApiResponse setPrecio(Context ctx) {
 		return mock.toBuilder()
-			.id(Integer.parseInt(ctx.pathParam("id")))
+			.id(Long.parseLong(ctx.pathParam("id")))
 			.precioAcordado(BigDecimal.TEN)
 			.build();
 	}
@@ -56,7 +56,7 @@ public class TransporteController {
 	 */
 	public ApiResponse inicio(Context ctx) {
 		return mock.toBuilder()
-			.id(Integer.parseInt(ctx.pathParam("id")))
+			.id(Long.parseLong(ctx.pathParam("id")))
 			.fechaInicio(ZonedDateTime.now(Dates.UTC))
 			.build();
 	}
@@ -66,7 +66,7 @@ public class TransporteController {
 	 */
 	public ApiResponse fin(Context ctx) {
 		return mock.toBuilder()
-			.id(Integer.parseInt(ctx.pathParam("id")))
+			.id(Long.parseLong(ctx.pathParam("id")))
 			.fechaInicio(ZonedDateTime.now(Dates.UTC).minusHours(1))
 			.fechaFin(ZonedDateTime.now(Dates.UTC))
 			.build();
@@ -77,7 +77,7 @@ public class TransporteController {
 	 */
 	public ApiResponse confirmarPago(Context ctx) {
 		return mock.toBuilder()
-			.id(Integer.parseInt(ctx.pathParam("id")))
+			.id(Long.parseLong(ctx.pathParam("id")))
 			.pagoConfirmado(true)
 			.build();
 	}
@@ -87,7 +87,7 @@ public class TransporteController {
 	 */
 	public ApiResponse confirmarEntrega(Context ctx) {
 		return mock.toBuilder()
-			.id(Integer.parseInt(ctx.pathParam("id")))
+			.id(Long.parseLong(ctx.pathParam("id")))
 			.entregaConfirmada(true)
 			.build();
 	}
